Tweak color handling and move BlogSection component to the right folder
