+++ packages/react-app/pages/index.jsx
@@ -7,7 +7,7 @@ import { getStats } from "../data/api/builder";
 import HeroSection from "../components/home/HeroSection";
 import ActivitySection from "../components/home/ActivitySection";
 import { getAllEvents } from "../data/api";
-import BlogSection from "../components/BlogSection";
+import BlogSection from "../components/home/BlogSection";
 import { fetchRecentPosts } from "../data/api/blog";
 const buildersToShow = ["fullstack", "frontend", "damageDealer", "advisor", "artist", "support"];
